backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
psycopg2-binary>=2.9.5,<3.0.0
django-redis>=5.2.0,<6.0.0
drf-yasg>=1.21.5,<2.0.0
requests>=2.28.2,<3.0.0
gunicorn>=21.2.0
jsonschema>=4.17.3,<5.0.0
